Fix: setup configuration for snort on linux

+++ scripts/install.sh
@@ -361,7 +361,12 @@ configure_snort_linux() {
     info_message "Configuring Snort"
     sed_alternative -i 's/output alert_fast: snort.alert.fast/output alert_fast: snort.alert/g' /etc/snort/snort.conf
     sed_alternative -i 's/# output alert_syslog: LOG_AUTH LOG_ALERT/output alert_syslog: LOG_AUTH LOG_ALERT/g' /etc/snort/snort.conf
-    maybe_sudo curl -o /etc/snort/rules/local.rules https://raw.githubusercontent.com/ADORSYS-GIS/wazuh-snort/refs/heads/main/scripts/windows/local.rules
+    if [! -f /etc/snort/rules/local.rules] then
+        maybe_sudo curl -o /etc/snort/rules/local.rules https://raw.githubusercontent.com/ADORSYS-GIS/wazuh-snort/refs/heads/main/scripts/windows/local.rules
+    else
+        error_message "File /etc/snort/rules/local.rules does not exist"
+        exit 1 
+    fi
     echo 'alert icmp any any -> any any (msg:"ICMP connection attempt:"; sid:1000010; rev:1;)' | maybe_sudo tee -a /etc/snort/rules/local.rules > /dev/null
 
     info_message "Downloading and configuring Snort rule files"
